# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "operationId": "UsersController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll_v1",
        "parameters": [
          {
            "name": "current",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        }
                      },
                      "required": [
                        "meta",
                        "result"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "operationId": "AdminController_getUsers_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_locked",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        }
                      },
                      "required": [
                        "meta",
                        "result"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUserDetail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      },
      "patch": {
        "operationId": "AdminController_updateUser_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserByAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteUser_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}/toggle-lock": {
      "patch": {
        "operationId": "AdminController_toggleUserLock_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}/role": {
      "patch": {
        "operationId": "AdminController_changeUserRole_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/admin/audit-logs": {
      "get": {
        "operationId": "AdminController_getAuditLogs_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "admin_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target_type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AdminActionLog"
                          }
                        }
                      },
                      "required": [
                        "meta",
                        "result"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "summary": "Nhật ký hành động quản trị (task #34).\n\nChỉ có đọc, và cố ý không có route xoá: một nhật ký xoá được thì không còn\nlà bằng chứng — mà đúng người muốn xoá nó là người có quyền admin.\n\n`GET` nên chính route này KHÔNG tự ghi vào nhật ký. Ghi cả lượt xem nhật ký\nsẽ làm bảng tự sinh ra dữ liệu về chính nó và đẩy hành động thật xuống dưới.",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/admin/withdrawals": {
      "get": {
        "operationId": "AdminController_getWithdrawals_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Withdrawal"
                          }
                        }
                      },
                      "required": [
                        "meta",
                        "result"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/withdrawals/{id}/approve": {
      "patch": {
        "operationId": "AdminController_approveWithdrawal_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Withdrawal"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/withdrawals/{id}/reject": {
      "patch": {
        "operationId": "AdminController_rejectWithdrawal_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Withdrawal"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/withdrawals/{id}/complete": {
      "patch": {
        "operationId": "AdminController_completeWithdrawal_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Withdrawal"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "summary": "Chặng thứ ba: tiền rời khỏi hệ thống sang `bank_external` sau khi admin đã\nchuyển khoản thật ngoài đời.\n\n`AdminService.completeWithdrawal` và `WithdrawalsService.complete` đều đã\ntồn tại và có test, nhưng KHÔNG có route nào gọi tới. Nghĩa là lệnh rút chỉ\nđi được tới `approved` rồi đứng đó vĩnh viễn, và tiền vẫn nằm trong\n`withdrawal_pending` — sổ cái nói người bán chưa được trả, dù thực tế đã\nchuyển. Thiếu đúng bảy dòng này.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/withdrawals": {
      "post": {
        "operationId": "WithdrawalsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWithdrawalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Withdrawal"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Withdrawals"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zoldify API",
    "description": "API của sàn mua bán đồ cũ Zoldify. Web (Next.js) và app (React Native) dùng chung hợp đồng này.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "phone_number": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "buyer",
              "seller",
              "admin",
              "moderator"
            ]
          },
          "avatar": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          }
        },
        "required": [
          "full_name",
          "email",
          "password"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "current": {
            "type": "number",
            "description": "Trang hiện tại, bắt đầu từ 1"
          },
          "pageSize": {
            "type": "number",
            "description": "Số bản ghi mỗi trang"
          },
          "pages": {
            "type": "number",
            "description": "Tổng số trang"
          },
          "total": {
            "type": "number",
            "description": "Tổng số bản ghi khớp điều kiện lọc"
          }
        },
        "required": [
          "current",
          "pageSize",
          "pages",
          "total"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "role": {
            "enum": [
              "buyer",
              "seller",
              "admin",
              "moderator"
            ],
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "is_locked": {
            "type": "boolean"
          },
          "last_seen": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "full_name",
          "email",
          "phone_number",
          "role",
          "avatar",
          "email_verified",
          "is_locked",
          "last_seen",
          "gender",
          "created_at"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
...
      "ChangeRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "buyer",
              "seller",
              "admin",
              "moderator"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "UpdateUserByAdminDto": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "phone_number": {
            "type": "string",
            "minLength": 8,
            "maxLength": 20
          },
          "avatar": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          }
        }
      },
...
      "CreateWithdrawalDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 10000
          },
          "bank_name": {
            "type": "string"
          },
          "bank_account": {
            "type": "string"
          },
          "bank_holder": {
            "type": "string"
          }
        },
        "required": [
          "amount",
          "bank_name",
          "bank_account",
          "bank_holder"
        ]
      }
    }
  }
}